ENH: reorganized project directory
* added two additional python packages:
  * QRCustomizations - customized classes
  * QRUtils - utils e.g. for html report generation
