+++ CMakeLists.txt
@@ -1,8 +1,7 @@
 cmake_minimum_required(VERSION 3.5.0)
 
-project(QuantitativeReporting)
-
 #-----------------------------------------------------------------------------
+set(EXTENSION_NAME QuantitativeReporting)
 set(EXTENSION_HOMEPAGE "http://wiki.slicer.org/slicerWiki/index.php/Documentation/Nightly/Modules/QuantitativeReporting")
 set(EXTENSION_CATEGORY "Informatics")
 set(EXTENSION_CONTRIBUTORS "Andrey Fedorov (SPL), Christian Herz (SPL), Csaba Pinter (Queen's), Andras Lasso (Queen's), Steve Pieper (SPL)")
@@ -13,28 +12,14 @@ set(EXTENSION_STATUS "Work in progress")
 set(EXTENSION_DEPENDS "SlicerDevelopmentToolbox DCMQI PETDICOMExtension")
 
 #-----------------------------------------------------------------------------
+
+set(MODULE_NAME QuantitativeReporting)
+
 find_package(Slicer REQUIRED)
 include(${Slicer_USE_FILE})
 
-#-----------------------------------------------------------------------------
-set(MODULE_NAME QuantitativeReporting)
-set(MODULE_TITLE ${MODULE_NAME})
-
-set(MODULE_PYTHON_RESOURCES
-  Resources/Icons/${MODULE_NAME}.png
-  )
-
-slicerMacroBuildScriptedModule(
-  NAME  ${MODULE_NAME}
-  SCRIPTS  ${MODULE_NAME}.py
-  RESOURCES ${MODULE_PYTHON_RESOURCES}
-  )
-
-string(TOUPPER ${MODULE_NAME} MODULE_NAME_UPPER)
-
-#-----------------------------------------------------------------------------
-
 add_subdirectory(DICOMPlugins)
+add_subdirectory(QuantitativeReporting)
 
 #-----------------------------------------------------------------------------
