+++ QuantitativeReporting/Resources/Icons/ReportingLogo.ai
